pass document instead of editor to getVariables()
VS Code can save documents that are not the active editor (i.e. mass
search/replace). Since the active editor was being used to determine
the file info for replacements on save, files were getting updated
with the wrong file info. This fixes the problem by passing the document
object instead of the editor object to the getVariables() function. This
way, the variable values are derived from the document being saved
instead of the active editor.

Fixes #21